requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0